TN-PCFG/requirement.txt
absl-py==0.11.0
aiohttp==3.7.4
allennlp==2.2.0
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==20.3.0
blessed==1.17.6
blessings==1.7
blis==0.7.4
boto3==1.17.43
botocore==1.20.43
cached-property==1.5.2
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
configparser==5.0.2
cvxpy==1.1.12
cycler==0.10.0
cymem==2.0.5
dataclasses==0.6
docker-pycreds==0.4.0
easydict==1.9
ecos==2.0.7.post1
filelock==3.0.12
fsspec==0.8.7
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.1
google-auth-oauthlib==0.4.3
gpustat==0.6.0
gql==2.0.0
graphql-core==2.3.2
grpcio==1.36.1
h5py==3.2.1
hydra-core==1.1.0.dev5
idna==2.10
importlib-metadata==3.4.0
importlib-resources==5.1.0
iniconfig==1.1.1
inquirer==2.7.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonnet==0.17.0
jsonpickle==2.0.0
kiwisolver==1.3.1
lmdb==1.1.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
mkl-fft==1.2.1
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.7.0
Mosek==9.2.40
multidict==5.1.0
murmurhash==1.0.5
nltk==3.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.0.dev22
opt-einsum==3.3.0
osqp==0.6.2.post0
overrides==3.1.0
packaging==20.9
pathtools==0.1.2
pathy==0.4.0
Pillow @ file:///tmp/build/80754af9/pillow_1609786740831/work
pluggy==0.13.1
preshed==3.0.5
prettytable==2.0.0
promise==2.3
protobuf==3.15.5
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-lightning==1.2.2
PyYAML==5.3.1
qdldl==0.1.5.post0
readchar==2.0.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rich==9.13.0
rsa==4.7.2
Rx==1.6.1
s3transfer==0.3.6
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.2
scs==2.1.3
sentencepiece==0.1.95
sentry-sdk==1.0.0
shortuuid==1.0.1
six @ file:///tmp/build/80754af9/six_1605205313296/work
smart-open==3.0.0
smmap==3.0.5
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
thinc==8.0.2
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
torch==1.6.0
torch-struct==0.5
torchaudio==0.7.0
torchvision==0.8.0
tqdm==4.57.0
transformers==4.3.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wandb==0.10.21
wasabi==0.8.2
watchdog==2.0.2
wcwidth==0.2.5
Werkzeug==1.0.1
yarl==1.6.3
zipp==3.4.0
triton==2.0.0

